Feat/assets 70 scheduling

+++ apps/ranking-service/src/main/java/io/antcamp/rankingservice/application/dto/RankingResult.java
@@ -1,12 +1,10 @@
-// application/dto/RankingResult.java
 package io.antcamp.rankingservice.application.dto;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 public record RankingResult(
         UUID userId,
-        BigDecimal totalAsset,
+        Double totalAsset,
         long rank
 ) {
 }
